Replaced 32bit pointers with 64 bit pointers
BOOL CALLBACK -> INT_PTR CALLBACk
GetWindowLong -> GetWindowLongPtr
SetWindowLong -> SetWindowLongPtr
DWL_USER -> DWLP_USER
GWL_USERDATA -> GWLP_USERDATA
GWL_HWNDPARENT -> GWLP_HWNDPARENT
long FAR PASCAL -> LRESULT FAR PASCAL
UINT (pointer) -> size_t
_findfirst and _findnext return intptr_t

+++ Src/Orbiter/Body.cpp
@@ -184,7 +184,7 @@ void Body::UnregisterVisual ()
 	hVis = NULL;
 }
 
-void Body::BroadcastVisMsg (DWORD msg, UINT content)
+void Body::BroadcastVisMsg (DWORD msg, size_t content)
 {
 	oapi::GraphicsClient *gc;
 	if (hVis && (gc = g_pOrbiter->GetGraphicsClient())) {
